+++ lab_3/asymmetric_keys/public.pem
@@ -1,9 +1,9 @@
 -----BEGIN PUBLIC KEY-----
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAnwyRptEGNrzKpUbRTThf
-RguPF+R8juZOykWwTRCkGhW0TgtKSufith5LW6Vh35hEiW2Bm+MV/TJyNHas8qvp
-RZVi6oaMGRXSZEHMIInNP9wRjZ6SOBENOfTwySysCkYRcQ2HoykyecbB1+i1GIPV
-ygQh3DV5p5bL/jfJEP4acrbALN13R8hMDumpciPAGnh5PVXEWheWf/q94U6gWJef
-b3sVSlVqaVrcxxILy3IYMydIGOmju9Pxcxfw9c8LvdJ48aPZ9Om8ER66YPRuABQX
-Now0vm3EG+Tqn2MZ3nC9qEfKtyZY4VHff+3lPyW1exllUdMfaHwY4dqag76Avi2Q
-uwIDAQAB
+MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAtufWrfPLOEUdy2E8/vb+
+W9OWZ7WjzVpwIfHo1IsWWOL64kP3LYDgOkNcPzsu1b+EzUGok+s463+TyCi61bsP
+t44vJuiOAvx5KaEbbOouaDjH8INzZ04ldKgqu9dpL3g1zEZ5F1wlrBb68h7rvNmc
+JbOKvMrMeOmO6GTIMUK4sPirrvi3BYKOeGPejxcEqCoUyTF7V9AtLzJRG2+Z2j5a
+/OLj0tbeJ9j/8o7HqIITP+2YSEesYvEVFqGOxEpR0dBvVtMdsT8tSAel7hu+svNE
+6PnZeGqaZoJD6ws6jTuOZAAM9s7EggpyFVmLBrDq3GwMVtYc6+1y13pL9Pa2qwDJ
+nwIDAQAB
 -----END PUBLIC KEY-----
