+++ lab_3/file_key_work.py
@@ -1,6 +1,9 @@
 import json
 import logging
 
+from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography.hazmat.primitives import serialization
+
 logging.basicConfig(level=logging.INFO)
 
 
@@ -39,6 +42,55 @@ class FileWork:
         except Exception as e:
             logging.error(f"Error in key_nonce_deserializer (FileWork) - {e}")
 
+    def serialize_private_key(self, private_key: rsa.RSAPrivateKey) -> None:
+        """
+        Serialize private key
+        :param private_key: rsa.RSAPrivateKey - private key:
+        :return None:
+        """
+        try:
+            with open(self.path, 'wb') as file:
+                file.write(private_key.private_bytes(encoding=serialization.Encoding.PEM,
+                                                     format=serialization.PrivateFormat.TraditionalOpenSSL,
+                                                     encryption_algorithm=serialization.NoEncryption()))
+        except Exception as e:
+            logging.error(f"Error in serialize_private_key (FileWork) - {e}")
+
+    def serialize_public_key(self, public_key: rsa.RSAPublicKey) -> None:
+        """
+        Serialize public key
+        :param public_key: rsa.RSAPublicKey - public key:
+        :return None:
+        """
+        try:
+            with open(self.path, 'wb') as file:
+                file.write(public_key.public_bytes(encoding=serialization.Encoding.PEM,
+                                                   format=serialization.PublicFormat.SubjectPublicKeyInfo))
+        except Exception as e:
+            logging.error(f"Error in serialize_public_key (FileWork) - {e}")
+
+    def deserialize_private_key(self) -> rsa.RSAPrivateKey:
+        """
+        Deserialize private key and return it
+        :return rsa.RSAPrivateKey - deserialized private key:
+        """
+        try:
+            with open(self.path, 'rb') as file:
+                return serialization.load_pem_private_key(file.read(), password=None)
+        except Exception as e:
+            logging.error(f"Error in deserialize_private_key (FileWork) - {e}")
+
+    def deserialize_public_key(self) -> rsa.RSAPublicKey:
+        """
+        Deserialize public key and return it
+        :return rsa.RSAPublicKey - deserialized public key:
+        """
+        try:
+            with open(self.path, 'rb') as file:
+                return serialization.load_pem_public_key(file.read())
+        except Exception as e:
+            logging.error(f"Error in deserialize_public_key (FileWork) - {e}")
+
     def txt_reader(self, mode: str, encoding=None) -> str:
         """
         Reading txt file and return string from there
@@ -52,10 +104,10 @@ class FileWork:
         except Exception as e:
             logging.error(f"Error in txt_reader (FileWork) - {e}")
 
-    def txt_writer(self, string: str) -> None:
+    def txt_writer(self, string: bytes) -> None:
         """
         Write string to file
-        :param string: str - string that should be written
+        :param string: bytes - string that should be written
         :return None:
         """
         try:
